Impostato utilizzo pairwise per random routing, ma non abilitato. Pulizie varie in serverrewind e correzioni in result_utils

+++ system/utils/result_utils.py
@@ -23,9 +23,13 @@ import os
 def average_data(algorithm="", dataset="", goal="", times=10):
     test_acc = get_all_results_for_one_algo(algorithm, dataset, goal, times)
 
+    if len(test_acc) == 0:
+        return
+
     max_accurancy = []
     for i in range(times):
-        max_accurancy.append(test_acc[i].max())
+        if i in test_acc:
+            max_accurancy.append(test_acc[i].max())
 
     print("std for best accurancy:", np.std(max_accurancy))
     print("mean for best accurancy:", np.mean(max_accurancy))
@@ -42,13 +46,17 @@ def get_all_results_for_one_algo(algorithm="", dataset="", goal="", times=10):
 
 
 def read_data_then_delete(file_name, delete=False):
+    if not os.path.exists("../results"):
+        os.makedirs("../results")
     file_path = "../results/" + file_name + ".h5"
+    rs_test_acc = []
 
-    with h5py.File(file_path, 'r') as hf:
-        rs_test_acc = np.array(hf.get('rs_test_acc'))
+    if os.path.exists(file_path):
+        with h5py.File(file_path, 'r') as hf:
+            rs_test_acc = np.array(hf.get('rs_test_acc'))
 
-    if delete:
-        os.remove(file_path)
-    print("Length: ", len(rs_test_acc))
+        if delete:
+            os.remove(file_path)
+        print("Length: ", len(rs_test_acc))
 
     return rs_test_acc
